requirements.txt

fastapi
uvicorn[standard]
pydantic

pydantic_ai
openai
python-dotenv
langgraph

pandas
numpy
scikit-learn
lime